Basic functionality, will only return json
