--- views/index.pug
@@ -1,23 +1,0 @@
-doctype html
-html(lang="en")
-    head
-        meta(charset="UTF-8")
-        meta(id="viewport" name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=2, viewport-fit=cover")
-        meta(name="theme-color" content="#ffffff")
-        meta(name="mobile-web-app-capable" content="yes")
-        link(href='/assets/style.css', rel='stylesheet', type='text/css', media='screen')
-        title= service_name
-    body
-        noscript
-            meta(http-equiv=refresh content='0; url="/assets/noscript.html"')
-        div(class="main search")
-            h1= service_name
-            p Browse Instagram content without logging in
-            form(role="search" action="/search")
-                input(type="search" id="site-search" name="q" minlength="1" maxlength="31" size="50" required autocomplete="off" spellcheck=false autocorrect=off placeholder="Search profiles, tags, and locations")
-                button <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-search" viewBox="0 0 16 16"><path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z"/></svg>
-        div(id="loading")
-            div(class="lds-spinner" style="top: 50%; transform: translateY(-50%); overflow: hidden")
-                | <div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div>
-        div(class="page")
-        script(defer src="/assets/Main.js")
