--- views/post.pug
@@ -1,68 +1,0 @@
-doctype html
-html(lang="en")
-  head
-    title= title
-    link(rel="icon" type="image/x-icon" href=`/user/${username}/thumb_pfp`)
-    link(rel="prefetch" href="/assets/loadPosts.js" as="script" crossorigin)
-    link(rel="modulepreload" href="https://cdn.jsdelivr.net/npm/ky@0.30.0/distribution/index.min.js" as="script" crossorigin)
-    link(rel="preload" href="/assets/postMeta.js" as="script" crossorigin)
-    link(href="/assets/js/splidejs/dist/css/themes/splide-default.min.css" rel="stylesheet")
-    meta(id="viewport" name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=2, viewport-fit=cover")
-    meta(name="theme-color" content="#ffffff")
-    meta(name="mobile-web-app-capable" content="yes")
-    meta(charset="UTF-8")
-    link(href='/assets/style.css', rel='stylesheet', type='text/css', media='screen')
-  
-  body
-    noscript
-      meta(http-equiv=refresh content='0; url="/assets/noscript.html"')
-    //- div(class="header")
-    //-   p(id="service_name")= service_name
-    //- WHERE IN THE WORLD EASTER EGG
-  div(class="main")
-    form(role="search" action="/search")
-      input(type="search" id="site-search" value="" name="q" minlength="1" maxlength="50" size="30" required autocomplete="off" spellcheck=false autocorrect=off placeholder="Search profiles, tags, and locations")
-      button <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-search" viewBox="0 0 16 16"><path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z"/></svg>
-  div(class="img_holder")
-    section(id="image-carousel" class="splide" aria-labelledby="carousel-heading")
-      div(class="splide__track")
-        div(class="splide__list")
-
-  div(class="main")
-    p(class="caption") 
-      a(href=`/user/${username}/` class="post_meta")
-        img(src=`/user/${username}/thumb_pfp` id="post_pfp" onmouseover=`this.src="/user/${username}/pfp"` onmouseout=`this.src="/user/${username}/thumb_pfp"` width=32 height=32 alt=`${username}'s Profile Picture`)
-        span(class="username_caption")= username
-        span(id="verified")
-          i(class=`${is_verified} `)
-      span(id="caption")  #{caption}
-    a(href="#" class="likes post_meta")
-      p(class="likes post_meta")= like_count
-    a(href="#" class="date right")
-      time(class="date right")= taken_at
-    details(id="details")
-      summary(class="comment post_meta") #{comment_count}
-      div(class="container")
-        div(class="lds-spinner first-load")
-        | <div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div>
-      div(class="page-load-status")
-        div(class="lds-spinner infinite-scroll-request")
-          | <div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div>
-
-  div(id="loading")
-      div(class="lds-spinner" style="top: 50%; transform: translateY(-50%); overflow: hidden")
-        | <div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div>
-  div(class="page")
-      script.
-        // Necessary for external javascript to work
-        var carousel_media_count = "#{carousel_media_count}";
-        var username = "#{username}";
-        var is_video = #{is_video};
-        var comment_count = "#{comment_count}"
-        var taken_at = "#{taken_at}";
-        var like_count = "#{like_count}";
-      script(defer src="/assets/Main.js")
-      script(async src="/assets/js/splidejs/dist/js/splide.min.js")
-      script(async type="module" src="/assets/postMeta.js")
-      script(async src="/assets/js/infinite-scroll.pkgd.min.js")
-      script(async src="/assets/js/lazysizes.min.js")
