--- views/search.pug
@@ -1,52 +1,0 @@
-doctype html
-html(lang="en")
-    head
-        meta(charset="UTF-8")
-        meta(id="viewport" name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=2, viewport-fit=cover")
-        meta(name="theme-color" content="#ffffff")
-        meta(name="mobile-web-app-capable" content="yes")
-        link(href='/assets/style.css', rel='stylesheet', type='text/css', media='screen')
-        title= service_name
-    body
-        noscript
-            meta(http-equiv=refresh content='0; url="/assets/noscript.html"')
-        div(class="main search")
-            form(role="search" action="/search")
-                    input(type="search" id="site-search" value="" name="q" minlength="1" maxlength="50" size="30" required autocomplete="off" spellcheck=false autocorrect=off placeholder="Search profiles, tags, and locations")
-                    button <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-search" viewBox="0 0 16 16"><path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z"/></svg>
-            ul(class="search")
-                each user in users
-                    - let verified = "";
-                    - if(user.user.is_verified) verified = "bi bi-patch-check-fill"
-                        li(class="search")
-                            div(class="search_container")
-                                a(href=`/user/${user.user.username}` class="post_meta")
-                                        img(data-src=`/user/${user.user.username}/thumb_pfp` onmouseover=`this.src="/user/${user.user.username}/pfp"` id="pfp" class="lazyload" onmouseout=`this.src="/user/${user.user.username}/thumb_pfp?preview=true"` width=96 height=96 alt=`${user.user.username}'s Profile Picture`)
-                                        div
-                                        span(class="post_meta") #{user.user.username} 
-                                            span(id="verified")
-                                                span(class=`${verified}`)
-                                        p(class="full_name_search")= user.user.full_name 
-                        
-        div(id="loading")
-            div(class="lds-spinner" style="top: 50%; transform: translateY(-50%); overflow: hidden")
-                | <div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div>
-        div(class="page")
-        script.
-            const input = document.querySelector("input");
-            const button = document.querySelector("button");
-            const query = new URL(window.location.href).searchParams.get("q");
-            input.value = query;
-            button.disabled = true; 
-            input.addEventListener('input', function () {
-                if(this.value === query) button.disabled = true; else button.disabled = false; 
-            });
-            const ul = document.querySelector("ul");
-            if(ul.childElementCount < 1) {
-                const text = document.createTextNode(`No results found for ${query}`)
-                const p = document.createElement("p");
-                p.appendChild(text);
-                ul.appendChild(p);
-            }
-        script(defer src="/assets/Main.js")
-        script(async src="/assets/js/lazysizes.min.js")
